# docs/swagger.yaml
basePath: /api/v1
definitions:
  presentation.GetAccountDetailsRequest:
    properties:
      users:
        items:
          type: string
        type: array
    type: object
  presentation.GetAccountDetailsResponse:
    properties:
      message:
        type: string
      result:
        type: boolean
      userDetails:
        items:
          $ref: '#/definitions/presentation.UserDetails'
        type: array
    type: object
  presentation.UserDetails:
    properties:
      company:
        type: string
      followers:
        type: integer
      login:
        type: string
      name:
        type: string
      publicRepos:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: apolinario.torresjr@gmail.com
    name: Pol Torres
  description: A service that provides user account details
  title: Account Detail APIs
  version: 1.0.0
paths:
  /users/{source}:
    post:
      consumes:
      - application/json
      description: Get details of all provided github users
      parameters:
      - description: List of users
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/presentation.GetAccountDetailsRequest'
      - description: Account source
        in: path
        name: source
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presentation.GetAccountDetailsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/presentation.GetAccountDetailsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/presentation.GetAccountDetailsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presentation.GetAccountDetailsResponse'
      summary: Get details of all provided github users
      tags:
      - user
swagger: "2.0"